+++ iothub_client/tests/iothubmessage_ut/main.c
@@ -7,5 +7,5 @@ int main(void)
 {
     size_t failedTestCount = 0;
     RUN_TEST_SUITE(iothubmessage_ut, failedTestCount);
-    return failedTestCount;
+    return (int)failedTestCount;
 }
